Fix seek to frame index out of bounds error message in GifInfoHandle

+++ android-gif-drawable/src/main/java/pl/droidsonroids/gif/GifInfoHandle.java
@@ -367,7 +367,7 @@ final class GifInfoHandle {
 	}
 
 	private void throwIfFrameIndexOutOfBounds(@IntRange(from = 0) final int index) {
-		final float numberOfFrames = getNumberOfFrames(gifInfoPtr);
+		final int numberOfFrames = getNumberOfFrames(gifInfoPtr);
 		if (index < 0 || index >= numberOfFrames) {
 			throw new IndexOutOfBoundsException("Frame index is not in range <0;" + numberOfFrames + '>');
 		}
